+++ base/exports.jl
@@ -753,6 +753,7 @@ export
     filter,
     foldl,
     foldr,
+    foreach,
     get,
     get!,
     getindex,
@@ -911,12 +912,9 @@ export
 
 # bigfloat & precision
     precision,
-    get_bigfloat_precision,
-    set_bigfloat_precision,
-    with_bigfloat_precision,
-    get_rounding,
-    set_rounding,
-    with_rounding,
+    rounding,
+    setprecision,
+    setrounding,
     get_zero_subnormals,
     set_zero_subnormals,
 
@@ -1084,6 +1082,7 @@ export
     isgeneric,
     isinteractive,
     less,
+    method_exists,
     methods,
     methodswith,
     module_name,
@@ -1300,6 +1299,7 @@ export
     tempname,
     touch,
     uperm,
+    walkdir,
 
 # external processes ## TODO: whittle down these exports.
     detach,